+++ lib/app/page/hrm/letter_myself_details/view/letter_myself_details_screen.dart
@@ -682,10 +682,10 @@ class LetterMyselfDetailsScreen extends GetView<LetterMyselfDetailsController> {
                             )
                           : controller.timeController.text,
                       period: controller.dayController.text == ""
-                          ? int.parse(controller
+                          ? double.parse(controller
                               .detailsSingle.value.rData!.period
                               .toString())
-                          : int.parse(controller.dayController.text),
+                          : double.parse(controller.dayController.text),
                       address: controller.addressController.text == ""
                           ? controller.detailsSingle.value.rData!.address
                               .toString()
@@ -732,10 +732,10 @@ class LetterMyselfDetailsScreen extends GetView<LetterMyselfDetailsController> {
                             )
                           : controller.timeController.text,
                       period: controller.dayController.text == ""
-                          ? int.parse(controller
+                          ? double.parse(controller
                               .detailsSingle.value.rData!.period
                               .toString())
-                          : int.parse(controller.dayController.text),
+                          : double.parse(controller.dayController.text),
                       address: controller.addressController.text == ""
                           ? controller.detailsSingle.value.rData!.address
                               .toString()
